obsidian-remote-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian-remote-mcp"
version = "0.1.0"
description = "Remote MCP server exposing Obsidian-like tools over FastMCP"
readme = "README.md"
authors = [{ name = "Fast Obsidian MCP", email = "oss@example.com" }]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.1.3",
    "fastapi>=0.110.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.6.0",
    "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "black>=23.12",
    "ruff>=0.3.4",
    "types-PyYAML>=6.0.12",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "S"]
ignore = ["S101", "S311"]

[tool.ruff.lint.isort]
known-first-party = ["obsidian_remote_mcp"]

[tool.pytest.ini_options]
addopts = "-ra --cov=obsidian_remote_mcp --cov-report=term-missing"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
strict = true
packages = ["obsidian_remote_mcp"]

[[tool.mypy.overrides]]
module = ["fastapi", "starlette.*", "fastmcp"]
ignore_missing_imports = true

[tool.coverage.run]
branch = true
source = ["obsidian_remote_mcp"]

[tool.coverage.report]
show_missing = true
skip_covered = true
